\name{interest-package}
\alias{interest-package}
\alias{interest}
\docType{package}
\title{
\packageTitle{interest}
}
\description{
\packageDescription{interest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{interest}
\packageIndices{interest}

Todo: description of the package
}
\author{
\packageAuthor{interest}

Maintainer: \packageMaintainer{interest}
}
\references{
F. Mercurio: Interest Rates
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
